ignore compiler folders
I'm implementing a new build process which includes compiling all the code from ES6 to ES5 and compress it. For this we will need to clone the traceur and closure compilers from Google. Here I just kicked out all this useless junk and added an entry for the comiler folders.