switch from Google Chrome to Chromium
Because Google never publicly named its browser engine (Blink and V8 as a bundle) and also not provides any unique identifier, we can never say "okay we are actually in a 'Blink/V8' engine". So I named the engine just Google Chrome and took "Chrome/" as an identifier form the user agent. But now Opera also uses the chromium source to build their browser, so I renamed the engine to Chromium to summarize all this browsers (Google Chrome, direct Chromium build, Opera). Identifier is still "Chrome/".